work8

+++ work8/index.js
@@ -1,7 +1,7 @@
 function calc(func){
     let result = document.getElementById('result')
-    let num1 = parseInt(document.getElementById('num1').value)
-    let num2 = parseInt(document.getElementById('num2').value)
+    let num1 = parseFloat(document.getElementById('num1').value)
+    let num2 = parseFloat(document.getElementById('num2').value)
     if(isNaN(num1) || isNaN(num2)){
         alert('请输入数字')
         return false
